+++ Packages/UGF.Types/Runtime/TypesUtility.cs
@@ -8,7 +8,7 @@ namespace UGF.Types.Runtime
     /// <summary>
     /// Provides utilities to work with types.
     /// </summary>
-    public static class TypesUtility
+    public static partial class TypesUtility
     {
         /// <summary>
         /// Gets enumerable through the all loaded types.
@@ -30,16 +30,13 @@ namespace UGF.Types.Runtime
         /// <param name="results">The collection to add found types.</param>
         /// <param name="identifierType">The identifier type of the type defines.</param>
         /// <param name="assembly">The assembly to search.</param>
-        /// <param name="inherit">Determines whether to search in inheritance chain to find the attribute.</param>
-        public static void GetTypes(ICollection<Type> results, Type identifierType = null, Assembly assembly = null, bool inherit = true)
+        public static void GetTypes(ICollection<Type> results, Type identifierType = null, Assembly assembly = null)
         {
             if (results == null) throw new ArgumentNullException(nameof(results));
 
-            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, inherit))
+            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, false))
             {
-                var attribute = type.GetCustomAttribute<TypeIdentifierAttribute>();
-
-                if (identifierType == null || attribute.IdentifierType == identifierType)
+                if (identifierType == null || TryGetIdentifierAttribute(type, identifierType, out _))
                 {
                     results.Add(type);
                 }
@@ -58,14 +55,13 @@ namespace UGF.Types.Runtime
         /// <param name="results">The collection to add found defines.</param>
         /// <param name="identifierType">The identifier type of the type defines.</param>
         /// <param name="assembly">The assembly to search.</param>
-        /// <param name="inherit">Determines whether to search in inheritance chain to find the attribute.</param>
-        public static void GetTypeDefines(ICollection<ITypeDefine> results, Type identifierType = null, Assembly assembly = null, bool inherit = true)
+        public static void GetTypeDefines(ICollection<ITypeDefine> results, Type identifierType = null, Assembly assembly = null)
         {
             if (results == null) throw new ArgumentNullException(nameof(results));
 
-            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, inherit))
+            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, false))
             {
-                var attribute = type.GetCustomAttribute<TypeDefineAttribute>();
+                var attribute = type.GetCustomAttribute<TypeDefineAttribute>(false);
 
                 if ((identifierType == null || attribute.IdentifierType == identifierType) && TryCreateType(type, out ITypeDefine define))
                 {
@@ -84,17 +80,14 @@ namespace UGF.Types.Runtime
         /// <param name="identifierType">The type of the identifier that provider supports.</param>
         /// <param name="assembly">The assembly to search.</param>
         /// <param name="includeDefines">Determines whether to include types from found type defines.</param>
-        /// <param name="inherit">Determines whether to search in inheritance chain to find the attribute.</param>
-        public static void GetTypes(ITypeProvider provider, Type identifierType, Assembly assembly = null, bool includeDefines = true, bool inherit = true)
+        public static void GetTypes(ITypeProvider provider, Type identifierType, Assembly assembly = null, bool includeDefines = true)
         {
             if (provider == null) throw new ArgumentNullException(nameof(provider));
             if (identifierType == null) throw new ArgumentNullException(nameof(identifierType));
 
-            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, inherit))
+            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, false))
             {
-                var attribute = type.GetCustomAttribute<TypeIdentifierAttribute>();
-
-                if (attribute.IdentifierType == identifierType)
+                if (TryGetIdentifierAttribute(type, identifierType, out _))
                 {
                     provider.TryAdd(type);
                 }
@@ -102,9 +95,9 @@ namespace UGF.Types.Runtime
 
             if (includeDefines)
             {
-                foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, inherit))
+                foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, false))
                 {
-                    var attribute = type.GetCustomAttribute<TypeDefineAttribute>();
+                    var attribute = type.GetCustomAttribute<TypeDefineAttribute>(false);
 
                     if (attribute.IdentifierType == identifierType && TryCreateType(type, out ITypeDefine define))
                     {
@@ -123,12 +116,11 @@ namespace UGF.Types.Runtime
         /// <param name="provider">The type provider to register.</param>
         /// <param name="assembly">The assembly to search.</param>
         /// <param name="includeDefines">Determines whether to include types from found type defines.</param>
-        /// <param name="inherit">Determines whether to search in inheritance chain to find the attribute.</param>
-        public static void GetTypes<T>(ITypeProvider<T> provider, Assembly assembly = null, bool includeDefines = true, bool inherit = true)
+        public static void GetTypes<T>(ITypeProvider<T> provider, Assembly assembly = null, bool includeDefines = true)
         {
             if (provider == null) throw new ArgumentNullException(nameof(provider));
 
-            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, inherit))
+            foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeIdentifierAttribute>(assembly, false))
             {
                 if (TryGetIdentifierFromType(type, out T identifier))
                 {
@@ -138,9 +130,9 @@ namespace UGF.Types.Runtime
 
             if (includeDefines)
             {
-                foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, inherit))
+                foreach (Type type in AssemblyUtility.GetBrowsableTypes<TypeDefineAttribute>(assembly, false))
                 {
-                    var attribute = type.GetCustomAttribute<TypeDefineAttribute>();
+                    var attribute = type.GetCustomAttribute<TypeDefineAttribute>(false);
 
                     if (attribute.IdentifierType == typeof(T) && TryCreateType(type, out ITypeDefine<T> define))
                     {
@@ -159,9 +151,9 @@ namespace UGF.Types.Runtime
         {
             if (type == null) throw new ArgumentNullException(nameof(type));
 
-            if (type.GetCustomAttribute<TypeIdentifierAttribute>() is ITypeIdentifierAttribute<T> attribute)
+            if (TryGetIdentifierAttribute(type, typeof(T), out TypeIdentifierAttribute attribute) && attribute is ITypeIdentifierAttribute<T> typeIdentifierAttribute)
             {
-                identifier = attribute.Identifier;
+                identifier = typeIdentifierAttribute.Identifier;
                 return true;
             }
 
@@ -173,20 +165,46 @@ namespace UGF.Types.Runtime
         /// Tries to get type identifier from the specified type that contains identifier attribute.
         /// </summary>
         /// <param name="type">The target type.</param>
+        /// <param name="identifierType">The type of the identifier.</param>
         /// <param name="identifier">The found identifier.</param>
-        public static bool TryGetIdentifierFromType(Type type, out object identifier)
+        public static bool TryGetIdentifierFromType(Type type, Type identifierType, out object identifier)
         {
             if (type == null) throw new ArgumentNullException(nameof(type));
+            if (identifierType == null) throw new ArgumentNullException(nameof(identifierType));
 
-            var attribute = type.GetCustomAttribute<TypeIdentifierAttribute>();
-
-            if (attribute != null)
+            if (TryGetIdentifierAttribute(type, identifierType, out TypeIdentifierAttribute attribute))
             {
                 identifier = attribute.GetIdentifier();
                 return true;
             }
 
             identifier = null;
+            return false;
+        }
+
+        /// <summary>
+        /// Tries to get identifier type attribute from the specified type that support specified identifier type.
+        /// </summary>
+        /// <param name="type">The target type.</param>
+        /// <param name="identifierType">The type of the identifier.</param>
+        /// <param name="attribute">The found attribute.</param>
+        public static bool TryGetIdentifierAttribute(Type type, Type identifierType, out TypeIdentifierAttribute attribute)
+        {
+            if (type == null) throw new ArgumentNullException(nameof(type));
+            if (identifierType == null) throw new ArgumentNullException(nameof(identifierType));
+
+            object[] attributes = type.GetCustomAttributes(false);
+
+            for (int i = 0; i < attributes.Length; i++)
+            {
+                if (attributes[i] is TypeIdentifierAttribute typeIdentifierAttribute && typeIdentifierAttribute.IdentifierType == identifierType)
+                {
+                    attribute = typeIdentifierAttribute;
+                    return true;
+                }
+            }
+
+            attribute = null;
             return false;
         }
 
@@ -221,6 +239,53 @@ namespace UGF.Types.Runtime
             try
             {
                 result = (T)Activator.CreateInstance(type);
+            }
+            catch (Exception e)
+            {
+                result = default;
+                exception = e;
+                return false;
+            }
+
+            exception = null;
+            return true;
+        }
+
+        /// <summary>
+        /// Tries to create specified type.
+        /// <para>
+        /// The type must contains default constructor.
+        /// </para>
+        /// </summary>
+        /// <param name="type">The type to create.</param>
+        /// <param name="arguments">The constructor arguments.</param>
+        /// <param name="result">The created result.</param>
+        public static bool TryCreateType<T>(Type type, object[] arguments, out T result)
+        {
+            if (type == null) throw new ArgumentNullException(nameof(type));
+            if (arguments == null) throw new ArgumentNullException(nameof(arguments));
+
+            return TryCreateType(type, arguments, out result, out _);
+        }
+
+        /// <summary>
+        /// Tries to create specified type.
+        /// <para>
+        /// The type must contains default constructor.
+        /// </para>
+        /// </summary>
+        /// <param name="type">The type to create.</param>
+        /// <param name="arguments">The constructor arguments.</param>
+        /// <param name="result">The created result.</param>
+        /// <param name="exception">The exception that could occurs during creation.</param>
+        public static bool TryCreateType<T>(Type type, object[] arguments, out T result, out Exception exception)
+        {
+            if (type == null) throw new ArgumentNullException(nameof(type));
+            if (arguments == null) throw new ArgumentNullException(nameof(arguments));
+
+            try
+            {
+                result = (T)Activator.CreateInstance(type, arguments);
             }
             catch (Exception e)
             {